Move composite keys find unique tests to its own module
